Fleshed out the Pistol weapon. Player can now shoot by clicking about with the mouse. Pistol handles reload and bullet to chamber wait times correctly (as far as I can tell). And you _can_ run out of ammo.
git-svn-id: file:///desktop/SourceForge%20Backup/isugamedev/svnrepos/trunk@1702 84b32ba4-53c3-423c-be69-77cca6335494

+++ midworld/src/Weapon.h
@@ -41,9 +41,9 @@ namespace mw
        * Updates the state of this bullet based on the amount of time that has
        * passed.
        *
-       * @param dt   the change in time in microseconds
+       * @param dt   the change in time in seconds
        */
-      virtual void update(u64 dt) = 0;
+      virtual void update(float dt) = 0;
    };
 }
